KATA-1147: PID-related UI-redesign
Remove one type of availability completely from the system (access_application_rems_other) and modify the existing access_application_rems_ida to be 'access_application_rems'. This means that this availability accepts any kind of identifier to be delivered to Reetta Rems service. If this identifier field is left empty, package id is used for sending to Reetta. If the value is non-empty, there is no longer validation for the format of the identifier since it basically can be anything. However, if the value is identified to be IDA PID, then access_application_download_URL can be and is filled out automatically. If the value is not IDA PID, it will not have a download URL.